Revert "Remove ErosScans, add DoomicScans (#1003)"
This reverts commit 4bd5f7582ca7b11f36e47817ff25926e3ccc13d9.
