+++ web/scripts/website-metrics.ts
@@ -34,6 +34,9 @@ type Result = {
 }
 
 const expectedRedirectPatterns = new Map([
+    [ 'https://erosxcomic.xyz/', [ // REASON: The website uses redirects to rotating (top-level) domains (probably to avoid scraping or DMCA)
+        /^https:\/\/(erosxcomic|erosmanga)\.xyz\/$/,
+    ] ],
     [ 'https://holymanga.net/', [ // REASON: The website uses redirects to rotating (sub-)domains (probably to avoid scraping or DMCA)
         /^https:\/\/w+\d*\.holymanga\.net\/$/,
     ] ],/*
