--- web/src/engine/websites/DoomicScans.ts
@@ -1,21 +1,0 @@
-import { Tags } from '../Tags';
-import icon from './DoomicScans.webp';
-import { DecoratableMangaScraper } from '../providers/MangaPlugin';
-import * as MangaStream from './decorators/WordPressMangaStream';
-import * as Common from './decorators/Common';
-
-@MangaStream.MangaCSS(/^{origin}\/manga\/[^/]+\/$/)
-@MangaStream.MangasSinglePageCSS()
-@MangaStream.ChaptersSinglePageCSS('div#chapterlist ul li div.chbox:not(:has(.dt)) div.eph-num a') //Exclude chapters on TecnoScans
-@MangaStream.PagesSinglePageJS()
-@Common.ImageAjax()
-export default class extends DecoratableMangaScraper {
-
-    public constructor() {
-        super('doomicscans', 'Doomic Scans', 'https://doomcomic.xyz', Tags.Media.Manhwa, Tags.Media.Manhua, Tags.Language.English, Tags.Source.Scanlator);
-    }
-
-    public override get Icon() {
-        return icon;
-    }
-}
