+++ web/src/engine/websites/ErosScans.ts
@@ -1,0 +1,27 @@
+import { Tags } from '../Tags';
+import icon from './ErosScans.webp';
+import { DecoratableMangaScraper } from '../providers/MangaPlugin';
+import * as MangaStream from './decorators/WordPressMangaStream';
+import * as Common from './decorators/Common';
+import { FetchWindowScript } from '../platform/FetchProvider';
+
+@MangaStream.MangaCSS(/^https:\/\/(erosxcomic|erosmanga)\.xyz\/manga\/[^/]+\/$/)
+@MangaStream.MangasSinglePageCSS()
+@MangaStream.ChaptersSinglePageCSS('div#chapterlist ul li div.chbox:not(:has(.dt)) div.eph-num a') //Exclude chapters on TecnoScans
+@MangaStream.PagesSinglePageJS()
+@Common.ImageAjax()
+export default class extends DecoratableMangaScraper {
+
+    public constructor() {
+        super('erosscans', 'Eros Scans', 'https://erosmanga.xyz', Tags.Media.Manhwa, Tags.Media.Manhua, Tags.Language.English, Tags.Source.Scanlator, Tags.Accessibility.DomainRotation);
+    }
+
+    public override get Icon() {
+        return icon;
+    }
+
+    public override async Initialize(): Promise<void> {
+        this.URI.href = await FetchWindowScript(new Request(this.URI), `window.location.origin`, 1500);
+        console.log(`Assigned URL '${this.URI}' to ${this.Title}`);
+    }
+}
