+++ web/src/engine/websites/_index.ts
@@ -121,7 +121,6 @@ export { default as Dmzj } from './Dmzj';
 export { default as Dokuha } from './Dokuha';
 export { default as DomalFansub } from './DomalFansub';
 export { default as DongManManhua } from './DongManManhua';
-export { default as DoomicScans } from './DoomicScans';
 export { default as Doujin69 } from './Doujin69';
 export { default as DoujinDesu } from './DoujinDesu';
 export { default as DoujinHentai } from './DoujinHentai';
@@ -138,6 +137,7 @@ export { default as EighteenComic } from './EighteenComic';
 export { default as EightMuses } from './EightMuses';
 export { default as EightMusesXXX } from './EightMusesXXX';
 export { default as ElevenToon } from './ElevenToon';
+export { default as ErosScans } from './ErosScans';
 export { default as EuphoriaScan } from './EuphoriaScan';
 export { default as EvilFlowers } from './EvilFlowers';
 export { default as FayScans } from './FayScans';
